requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
pydantic==2.8.2
python-dotenv==1.0.0
supabase==2.7.4
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.4.2
pandas==2.2.2
Jinja2==3.1.3
textblob==0.18.0.post0
google-search-results==2.4.2
chromadb==0.4.24
redis==5.0.0

# Multi-Modal Analysis Dependencies
Pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
imageio==2.31.1
matplotlib==3.7.1
scipy==1.11.1
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2

# Causal Inference Libraries for Task 1.1
dowhy==0.11.1
econml==0.15.0
causal-learn==0.1.3.8
networkx==3.1
graphviz==0.20.1
